Allow .mid/.MID extension on basename
This changes the command line handling so if the user
specifies the full name of the MIDI file, the code will silently
drop the extension and use that as the base name. This makes
the tool friendlier for people using filename completion, as they
don't have to remember to remove the ".mid" manually.